Updating the trace plots
Hi @beverlytan I'm just trying to start this pull request as a way to ask if you could fix up some of the graphs here, let's see if this works!

If you check out the tutorial, the graphs for the trace plots of the models (the caterpillar ones) all have this dark shadow towards the bottom - it's cause they are screenshots that I took in a hurry. I was wondering if you'd be willing to replace the files with ones without those shadows (they could still be screencaps, just better screencaps than mine).

Thanks!